+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 2.1.1 (Next)
+**Packer**
+- Add `parse` method.
 
 ### 2.1.0
 **Helm**